% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credit_data}
\alias{credit_data}
\title{Credit Default Dataset}
\format{
A data frame with 500 rows and 4 variables:
\describe{
  \item{age}{Age of the customer in years}
  \item{income}{Annual income in dollars}
  \item{debt_ratio}{Debt-to-income ratio (0 to 1)}
  \item{default}{Binary indicator: 1 = default, 0 = no default}
}
}
\usage{
credit_data
}
\description{
A synthetic dataset containing information about credit customers and their
default status.
}
\examples{
data(credit_data)
head(credit_data)
table(credit_data$default)
}
\keyword{datasets}
