% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticR-package.R
\docType{package}
\name{logisticR-package}
\alias{logisticR}
\alias{logisticR-package}
\title{logisticR: Efficient Logistic Regression Implementation}
\description{
An efficient implementation of logistic regression using gradient descent. This package provides functions for fitting binary logistic regression models and making predictions.
}
\author{
\strong{Maintainer}: Jiadong Student \email{jiadong@umich.edu}

}
\keyword{internal}
