% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_cpp.R
\name{my_logistic_cpp}
\alias{my_logistic_cpp}
\title{Fit Logistic Regression with C++ Acceleration}
\usage{
my_logistic_cpp(
  X,
  y,
  learning_rate = 0.01,
  max_iter = 1000,
  tol = 1e-06,
  add_intercept = TRUE
)
}
\arguments{
\item{X}{Predictor matrix}

\item{y}{Binary response vector}

\item{learning_rate}{Learning rate, default 0.01}

\item{max_iter}{Maximum iterations, default 1000}

\item{tol}{Convergence tolerance, default 1e-6}

\item{add_intercept}{Add intercept, default TRUE}
}
\value{
A my_logistic object
}
\description{
Fit Logistic Regression with C++ Acceleration
}
\examples{
X <- matrix(rnorm(200), ncol = 2)
y <- rbinom(100, 1, 0.5)
model <- my_logistic_cpp(X, y)
}
