% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{print.my_logistic}
\alias{print.my_logistic}
\title{Print method for my_logistic}
\usage{
\method{print}{my_logistic}(x, ...)
}
\arguments{
\item{x}{A my_logistic object}

\item{...}{Additional arguments}
}
\description{
Print method for my_logistic
}
