% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.my_logistic}
\alias{summary.my_logistic}
\title{Summary Method for my_logistic Objects}
\usage{
\method{summary}{my_logistic}(object, ...)
}
\arguments{
\item{object}{An object of class "my_logistic"}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for my_logistic Objects
}
